Split server
