+++ Cargo.toml
@@ -5,4 +5,5 @@ members = [
   "propolis-cli",
   "propolis-client",
   "propolis-server",
+  "propolis-server-lib",
 ]
