+++ propolis-server-lib/src/initializer.rs
