+++ propolis-server-lib/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod config;
+mod initializer;
+mod serial;
+pub mod server;
