+++ propolis-server-lib/src/serial.rs
