+++ propolis-server-lib/src/server.rs
@@ -416,62 +416,3 @@ pub fn api() -> ApiDescription<Context> {
     api.register(instance_serial).unwrap();
     api
 }
-
-#[cfg(test)]
-mod tests {
-    use dropshot::{
-        ConfigDropshot, ConfigLogging,
-        ConfigLoggingLevel, HttpServer, HttpServerStarter
-    };
-    use reqwest;
-    use std::collections::BTreeMap;
-    use std::net::SocketAddr;
-    use crate::config::{Config, Device};
-    use crate::test_utils::{setup, TEST_IMAGE, TEST_BOOTROM};
-
-    async fn initialize_server() -> HttpServer<super::Context> {
-        setup().await;
-
-        let mut block_options = BTreeMap::new();
-        block_options.insert(
-            "disk".to_string(),
-            toml::value::Value::String(TEST_IMAGE.path().as_path().to_str().unwrap().to_string())
-        );
-        block_options.insert(
-            "pci-path".to_string(),
-            toml::value::Value::String("0.4.0".to_string()),
-        );
-        let block = Device {
-            driver: "pci-virtio-block".to_string(),
-            options: block_options,
-        };
-        let mut devices = BTreeMap::new();
-        devices.insert("block0".to_string(), block);
-
-        let config = Config::new(TEST_BOOTROM.path(), devices);
-        let context = super::Context::new(config);
-
-        let config_dropshot = ConfigDropshot {
-            bind_address: "127.0.0.1:0".parse().unwrap(),
-            ..Default::default()
-        };
-        let config_logging =
-            ConfigLogging::StderrTerminal { level: ConfigLoggingLevel::Info };
-        let log = config_logging
-            .to_logger("propolis-server")
-            .unwrap();
-        HttpServerStarter::new(&config_dropshot, super::api(), context, &log)
-            .unwrap()
-            .start()
-    }
-
-    #[tokio::test]
-    async fn test_simple() {
-        let server = initialize_server().await;
-        server.close().await.unwrap();
-
-        // TODO: Use server.local_addr() + propolis client
-
-        assert!(false);
-    }
-}
