+++ propolis-server-lib/tests/artifacts.rs
